UBUNTU: Move isofs to storage-core-modules udeb from fs-core-modules.
I know there are no fs/ modules in storage-core-modules at the moment,
but it seems to be the most appropriate out of the udebs currently in
d-i's cdrom initrd, and taking fs-core-modules out of the cdrom initrd
will save 1MB there. We identified this issue just before Feisty Herd 2,
but apparently forgot to do anything about it.

Signed-off-by: Colin Watson <cjwatson@ubuntu.com>
Signed-off-by: Ben Collins <bcollins@ubuntu.com>