UBUNTU: build/config: Build ide-core as a module
It used to be that we had this built-in so people could pass command line
args to it. Hopefully we'll have a work around for it in gutsy to pass
module params on the kernel command line. If not, I believe this module
isn't being used enough to be an issue anyway (move to libata). Not to
mention we handle the main option (all_generic_ide).

Signed-off-by: Ben Collins <bcollins@ubuntu.com>