UBUNTU: Refresh laptop lid status on resume
OriginalLocation: https://bugzilla.redhat.com/show_bug.cgi?id=243030
Bug: 125855

On resume we need to refresh the lid status as we will not get an event if
the lid opening was what triggered the suspend.
This manifests itself in users never getting a "lid open" event when a
suspend happens because of lid close on hardware that supports wake on
lid open. This makes userspace gets very confused indeed.
Patch attached forces a check of the lid status in the resume handler.

Signed-off-by: Richard Hughes <richard@hughsie.com>
Signed-off-by: Amit Kucheria <amit.kucheria@ubuntu.com>