UBUNTU: hostap: send events on data interface as well as master interface
The hostap driver provides two interfaces - a master interface for raw access
to the hardware, and a data interface that presents an ethernet-style device.
Currently it only sends wireless events (such as association) to the former,
but applications like network-manager expect the events to come from the
interface that they manipulated. In order to avoid breaking software that might
depend on the former, just send both.

Signed-off-by: Matthew Garrett <mjg59@srcf.ucam.org>