UBUNTU: acpi_scan_rsdp() breaks some PCs by not honouring ACPI specification
2007-09-24 TJ <linux@tjworld.net>
The ACPI specification states the first 20 bytes of the RSDP table
must have a checksum of 0 (ACPI 1.0b RSDP table is 20 bytes long).
The signature can appear in multiple memory locations so don't rely
on it as the sole proof of a valid table.
This fixes broken/disabled ACPI problems with Acer Travelmate C100
(and others) where the first signature match is accepted without
confirming the checksum.

Signed-off-by: Matthew Garrett <mjg59@srcf.ucam.org>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>