[PATCH]: Gutsy OHCI hang workaround for Huron
I'm submitting this outside of my ldom2622 tree because it
really isn't a sparc change :-)

This keeps the root hub port reset routine in the OHCI driver
from hanging if the chip never clears the reset bit in it's
register.  For some reason this is happening on P11 and later
Huron systems, and I am actively trying to find out why.

Regardless of the reason for the hardware getting stuck, the
change below makes sense because no code should loop forever
hanging the kernel when this kind of thing happens.  I intend
to push this upstream, or something very similar to it.

Also, all the necessary fixes for the NIU driver are in my
ldom2622 tree, so if you could do a quick pull I'd appreciate
it.

Thanks!

[USB] ohci: Do not hang the system if port reset does not complete.

Signed-off-by: David S. Miller <davem@davemloft.net>