nfsd4: probe callback channel only once
BugLink: https://bugs.launchpad.net/ubuntu/+source/linux/+bug/253004

Our callback code doesn't actually handle concurrent attempts to probe
the callback channel.  Some rethinking of the locking may be required.
However, we can also just move the callback probing to this case.  Since
this is the only time a client is "confirmed" (and since that can only
happen once in the lifetime of a client), this ensures we only probe
once.

Signed-off-by: J. Bruce Fields <bfields@citi.umich.edu>
Signed-off-by: Manoj Iyer <manoj.iyer@canonical.com>
Acked-by: Tim Gardner <tim.gardner@canonical.com>
Acked-by: Stefan.Bader <stefan.bader@canonical.com>