UBUNTU: SAUCE: make /dev/kmem a config option OriginalAuthor: Arjan van de Ven <arjan@linux.intel.com> OriginalLocation: http://lkml.org/lkml/diff/2008/2/10/328/1
This is a sauce patch for now 'cause its not clear when or if
it will make it upstream. See discussion at http://lkml.org/lkml/2008/2/10/316

This patch makes /dev/kmem a config option; /dev/kmem is VERY rarely
used, and when used, it's generally for no good (rootkits tend to be
the most common users). With this config option, users have the
choice to disable /dev/kmem, saving some size as well.

Signed-off-by: Arjan van de Ven <arjan@linux.intel.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>