UBUNTU: x86: Clear DF before calling signal handler
OriginalAuthor: Aurelien Jarno <aurelien@aurel32.net>
Bug: #199072

The Linux kernel currently does not clear the direction flag before
calling a signal handler, whereas the x86/x86-64 ABI requires that.
This become a real problem with gcc version 4.3, which assumes that
the direction flag is correctly cleared at the entry of a function.

This patches changes the setup_frame() functions to clear the
direction before entering the signal handler.

This patch was committed upstream as e40cd10ccff3d9fbffd57b93780bee4b7b9bff51.
It was modified to exclude the changes introduced by
053de044411111da00272d1b4e174e7dd743f499.

Signed-off-by: Aurelien Jarno <aurelien@aurel32.net>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>