UBUNTU: SAUCE: Add blacklist support to fix Belkin bluetooth dongle. Bug: #140511
The Belkin bluetooth dongle unfortunately shares the vendor and device id with
the network adapter which causes lockups whenever the bluetooth dongle is
inserted. The preferred way upstream would be a mor detailed match function
but this will probably take more time. For the time being this helps people
with the bluetooth adapter.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>