UBUNTU: Add fancontrol modules to powerpc64-smp debian installer OriginalAuthor: Colin Watson <cjwatson@canonical.com>
The attached patch adds a fancontrol-modules udeb for powerpc64-smp. d-i
needs to have these modules available in order to avoid aeroplane-engine
fan syndrome during installation, which is a frequent complaint. The
approach in this patch is a little awkward because it needs to generate
the udeb just for one flavour on one architecture, but I've build-tested
it on powerpc and it seems to do the right thing. The list of modules
for this udeb, as well as the udeb name and description, is taken
directly from Debian's linux-kernel-di-powerpc-2.6 package.

If this could go in for 8.04.1, I'd appreciate it; the fan noise seems
to be a major off-putting issue for powerpc users, and it's a fairly
non-invasive change.

Signed-off-by: Colin Watson <cjwatson@canonical.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>