UBUNTU: cx88: enable radio GPIO correctly OriginalAuthor:  Steven Toth <stoth@hauppauge.com> Bug: #209971
commit dcd8f5bca3782f180c028446710edc16ebce73f6 (linux-2.6.24.y)
Author: Steven Toth <stoth@hauppauge.com>
Date:   Thu Apr 24 20:52:42 2008 -0400

This patch fixes an issue on the HVR1300, where GPIO is blown away due to
the radio input being undefined, breaking the functionality of the DVB
demodulator and MPEG2 encoder used on the cx8802 mpeg TS port.

This is a minimal patch for 2.6.26 and the -stable series.  This must be
fixed a better way for 2.6.27.

Signed-off-by: Steven Toth <stoth@hauppauge.com>
Signed-off-by: Mauro Carvalho Chehab <mchehab@infradead.org>
Signed-off-by: Michael Krufky <mkrufky@linuxtv.org>
(cherry picked from commit 6b92b3bd7ac91b7e255541f4be9bfd55b12dae41)
Signed-off-by: Greg Kroah-Hartman <gregkh@suse.de>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>