UBUNTU: rpcb_getport_async in sunrpc can cause oops on Hardy OriginalAuthor: HIRANO Takahito <hiranotaka@zng.info> OriginalLocation: http://launchpadlibrarian.net/14045528/sunrpc-memcpy-fix.patch Bug: #224750
The rpcb_getport_async function in the sunrpc module copies larger
memory area than the allocated on Ubuntu Hardy.
Because of this problem, NFS users might encounter oops. See Bug #224750.

This bug is derived from Linux 2.6.24 in kernel.org.
It seems to be fixed on Linux 2.6.25 in kernel.org by the commit
86d61d8638ddf9cdf87df26c7fa69b2804425fbe.

Signed-off-by: HIRANO Takahito <hiranotaka@zng.info>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>