UBUNTU: b43: Workaround invalid bluetooth settings OriginalAuthor: Michael Buesch <mb@bu3sch.de> OriginalLocation: http://launchpadlibrarian.net/14435339/0001-b43_invalid_bluetooth Bug: #197959
This adds a workaround for invalid bluetooth SPROM settings
on ASUS PCI cards.

This will stop the microcode from poking with the BT GPIO line.
This fixes data transmission on this device, as the BT GPIO line
is used for something TX related on this device
(probably the power amplifier or the radio).

This patch was applied to mainline kernels with commit
1855ba7812dbd294fcfc083dc7d3b14d3b1f38db and was authored by
Michael Buesch <mb@bu3sch.de>.

Signed-off-by: Larry Finger <Larry.Finger@lwfinger.net>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>