UBUNTU: ssb: Fix IRQ vectors enable for early cards. OriginalAuthor: Larry Finger <Larry.Finger@lwfinger.net> OriginalLocation: http://launchpadlibrarian.net/14435341/0005-irqvecs_enable Bug: #197959
After the release of 2.6.24, an error was found in the setup for
IRQ vectors in the ssb driver. This error was fixed in the mainline
kernels with commit b63009b456c8d9abe684bdf8d4bd8f27eb040019, which
was authored by me. This bug prevents any data transmission by the
affected cards.

Signed-off-by: Larry Finger <Larry.Finger@lwfinger.net>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>