UBUNTU: ssb: Fix all-ones boardflags OriginalAuthor:  Larry Finger <Larry.Finger@lwfinger.net> OriginalLocation: http://launchpadlibrarian.net/14435353/0007-ssb_boardflags_fix Bug: #197959
In the SSB SPROM a field set to all ones means the value
is not defined in the SPROM.

In case of the boardflags, we need to set them to zero
to avoid confusing drivers. For certain cards, leaving
the flags all set to 1 will prevent the card from transmitting,
as has been seen for the PCI version of a BCM4301.

This patch was included in mainline kernels with commit
4503183aa32e6886400d82282292934fa64a81b0.

Signed-off-by: Larry Finger <Larry.Finger@lwfinger.net>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>