UBUNTU: b43legacy: fix hard crash when BCM4303 present. OriginalAuthor: Stefano Brivio <stefano.brivio@polimi.it> OriginalLocation: http://launchpadlibrarian.net/14435495/0002-b43legacy_hard_crash Bug: #192720
Upon driver load when a BCM4304 rev 2 card is present, the kernel undergoes
a hard crash with power off the only recovery. This crash is due to a missing
check for NULL pointers, which only occur for a limited set os devices.

The fix in mainline kernels was submitted as
commit 881400a20c3551e90eed1062cf0387fa686a2fd0 by Stefano Brivio
<stefano.brivio@polimi.it>.

Signed-off-by: Larry Finger <Larry.Finger@lwfinger.net>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>