UBUNTU: drivers: fix dma_get_required_mask Bug: #238118
commit e88a0c2ca81207a75afe5bbb8020541dabf606ac
Author: James Bottomley <James.Bottomley@HansenPartnership.com>
Date:   Sun Mar 9 11:57:56 2008 -0500

There's a bug in the current implementation of dma_get_required_mask()
where it ands the returned mask with the current device mask.  This
rather defeats the purpose if you're using the call to determine what
your mask should be (since you will at that time have the default
DMA_32BIT_MASK).  This bug results in any driver that uses this function
*always* getting a 32 bit mask, which is wrong.

Fix by removing the and with dev->dma_mask.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>
Cc: stable <stable@kernel.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@suse.de>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>