UBUNTU: SAUCE: e1000 checksum recheck after reset Bug: #60388
The e1000 adapter appears to be initialized in a variety of states. If
a checksum check fails, then reset the adapter and try again. This
patch (or some form of it) has been floating around in various upstream
bug reports, then refactored by me. I could not figure out who the
original author was. I tested on a ThinkPad T60 with docking station.

Signed-off-by: Tim Gardner <tim.gardner@canonical.com>