UBUNTU: SAUCE: Add the ability to whitelist systems to use ACPI reboot
OriginalAuthor: yakui.zhao@intel.com
OriginalLocation: http://bugzilla.kernel.org/show_bug.cgi?id=7299
Bug: #249296

Some systems can be rebooted with the ACPI method although the checks/flags
would not indicate this. This patch adds a whitelist mechanism to the ACPI
reboot code and adds the classmate PC.
To use the ACPI method it is still necessary to specify reboot=a on the
kernel command line.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>